feat(ably-grid-10): install and init spaces
