+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@ably-labs/react-hooks": "^2.0.7",
+    "@ably/spaces": "^0.2.0",
     "@ag-grid-community/client-side-row-model": "^29.2.0",
     "@ag-grid-community/react": "^29.2.0",
     "@ag-grid-community/styles": "^29.2.0",
@@ -19,6 +20,7 @@
     "@ag-grid-enterprise/range-selection": "^29.2.0",
     "@faker-js/faker": "^7.6.0",
     "@heroicons/react": "^1.0.6",
+    "ably": "^1.2.47",
     "jose": "^4.9.3",
     "nanoid": "^4.0.0",
     "random-words": "^1.2.0",
